% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.edgwas.R
\name{boot.edgwas}
\alias{boot.edgwas}
\title{Bootstrap for sensitivity analysis of EdGwas}
\usage{
boot.edgwas(x, y, rho = NULL, nfolds = 10, type.measure = "rand",
  ...)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format.}

\item{y}{response matrix, of dimension nobs x nouts. Quantitative for family="gaussian". For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; for a factor, the last level in alphabetical order is the target class). For "binomial" if y is presented as a vector, it will be coerced into a factor.}

\item{rho}{(Non-negative) optional user-supplied rho sequence; default is NULL, and EdGwas chooses its own sequence.}

\item{nfolds}{number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds=3.}

\item{type.measure}{loss to use for cross-validation. Currently one option; the default is type.measure="rand", which uses the rand index.}
}
\value{
Cluster associations. \item{clust}{returns a vector with group memberships}
}
\description{
Does bootstrap for EdGwas, produces a plot, and returns a value for rho.
}
