% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.edgwas.R
\name{cv.edgwas}
\alias{cv.edgwas}
\title{Cross-validation for edgwas}
\usage{
cv.edgwas(x, y, scores = TRUE, rho = NULL, nfolds = 10,
  type.measure = c("mse", "mae"), nrho = ifelse(is.null(rho), 40,
  length(rho)), logrho = FALSE, rho.min.ratio = 0.001)
}
\arguments{
\item{x}{Input matrix, of dimension nobs x nvars or nobs x nouts; each row is an observation vector. A matrix of polygenic scores (PSs) if \code{scores = TRUE} (default) and nvars = nouts. Can be in sparse matrix format.}

\item{y}{Quantitative response matrix, of dimension nobs x nouts.}

\item{scores}{Are PSs provided (default is TRUE) or should PSs be generated from x variable (FALSE).}

\item{rho}{(Non-negative) optional user-supplied rho sequence; default is \code{NULL}, and EdGwas chooses its own sequence.}

\item{nfolds}{Number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is \code{nfolds = 3}.}

\item{type.measure}{Loss to use for cross-validation. Currently two options; the default is \code{type.measure="mse"}, which uses the mean-squared error. \code{type.measure = "mae"} gives the mean absolute error.}

\item{nrho}{The number of rho values. Default is 40.}

\item{logrho}{Logical flag for log transformation of the rho sequence. Default is \code{logrho = FALSE}.}

\item{rho.min.ratio}{Smallest value for rho, as a fraction of rho.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero) - default is 10e-04.}
}
\value{
An object of class "cv.edgwas" is returned, which is a list with results of the cross-validation fit. \item{rho}{The values of rho used in the fits.} \item{cvm}{The mean cross-validated error - a vector of length length(rho).} \item{cvsd}{ The estimate of the standard error of cvm.} \item{cvup}{Upper curve = cvm + cvsd.} \item{cvlo}{Lower curve = cvm - cvsd.} \item{name}{A text string indicating type of measure (for plotting purposes).} \item{rho.min}{Value of rho that gives minimum cvm.} \item{rho.1se}{Smallest value of lambda such that the error is within 1 standard error of the minimum.} \item{edgwas.fit}{A fitted edgwas object for the full data.}
}
\description{
Does k-fold cross-validation for edgwas, produces multiple plots, and returns a value for rho.
}
\details{
The function runs edgwas nfolds+1 times; the first to get the rho sequence, and then the remainder to compute the fit with each of the folds omitted. The error is accumulated, and the average error and standard deviation over the folds is computed. Note that the results of cv.edgwas are random, since the folds are selected at random. Users can reduce this randomness by running cv.edgwas many times, and averaging the error curves.
}
\examples{
N <- 1000 #
q <- 10 #
p <- 100
set.seed(1)
x <- matrix(sample(0:2, N*p, replace=TRUE), nrow=N, ncol=p)
B <- matrix(0, nrow = p, ncol = q)
B[1, 1:2] <- 5
y <- x \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
###
pc <- cv.edgwas(x, y, scores = FALSE)
\dontrun{
plot(pc, 1)
plot(pc, 1, zoom = 10)
plot(pc, 2)
}

}
