% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.edgwas.R
\name{cvcompute}
\alias{cvcompute}
\title{Internal EdGwas functions.}
\usage{
cvcompute(outlist, rho, PS, y, nfolds, P, type.measure, logrho,
  rho.min.ratio)
}
\description{
Internal EdGwas functions.
}
\details{
These are not intended for use by users but may of interest to developers. cvcompute does cross-validation.
}
\keyword{internal}
