% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdGWAS-package.R
\docType{package}
\name{EdGWAS-package}
\alias{EdGWAS-package}
\title{EdGWAS:}
\description{
A package implementing methods for clustering traits that share some genetic component via polygenic scores (PSs). It fits a sparse precision matrix of the PSs via graphical lasso. The regularisation path is computed for the lasso penalty rho at a grid of values for the regularisation parameter rho. A cross-validation method for tuning the penalty rho is provided.
}
\details{
The EdGWAS package provides three categories of important functions:
simple fit, cross-validation and plots.
}
\section{fit}{

The function \code{\link{edgwas}} clusters outcome components (traits) that share some feature (genetic component) via polygenic scores (PSs). It fits a sparse precision matrix via graphical lasso. The regularisation path is computed for the lasso penalty at a grid of values for the regularisation parameter rho. The clusters are used to specify the structure of the error covariance matrix of a GLS model and the feasible GLS estimator is used for estimating the unknown parameters in a linear regression model with a certain unknown degree of correlation between the residuals.
}

\references{
glasso, glmnet
}
\author{
Ann-Sophie Buchardt \email{asbu@sund.ku.dk}\cr Maintainer: Ann-Sophie Buchardt
\email{asbu@sund.ku.dk}
}
