% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matplot.edgwas.R
\name{matplot.edgwas}
\alias{matplot.edgwas}
\title{Adjacency matrix plot for a cv.edgwas object}
\usage{
matplot.edgwas(x, s = c("rho.min", "rho.1se"), A = NULL,
  col = "darkgrey", bty = "n", axes = TRUE, frame.plot = FALSE,
  ...)
}
\arguments{
\item{x}{Fitted "cv.edgwas" object.}

\item{s}{Value(s) of the penalty parameters rho for which the corresponding adjacency matrix is plotted. Default is rho.min.}

\item{A}{True adjacency matrix (optional).}

\item{col}{Color(s) to fill or shade the non-zero entries with. The default is \code{col = "lightgrey"}.}

\item{bty}{A character string which determine the type of box which is drawn about plots. Default is "n" which suppresses the box.}

\item{axes}{A logical value indicating whether axes should be drawn on the plot.}

\item{frame.plot}{A logical value indicating whether a box should be drawn around the plot.}

\item{...}{Other paramters to be passed through to plotting functions.}
}
\description{
Adjacency matrix plot for a cv.edgwas object
}
\details{
Plots are produced, and nothing is returned.
}
