% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.edgwas.R
\name{plot.cv.edgwas}
\alias{plot.cv.edgwas}
\title{Plot the cross-validation curve produced by cv.edgwas}
\usage{
\method{plot}{cv.edgwas}(x, ...)
}
\arguments{
\item{x}{Fitted "cv.edgwas" object.}
}
\description{
Plots the cross-validation curve, and empirical 95% confidence band, as a function of the rho values used.
}
\details{
A plot is produced, and nothing is returned.
}
