% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.edgwas.R
\name{plot.edgwas}
\alias{plot.edgwas}
\title{Plot diagnostics for an edgwas object}
\usage{
\method{plot}{edgwas}(x, which = c(1L:2L), main = "", xlab = NULL,
  ylab = NULL, ask = prod(par("mfcol")) < length(which) &&
  dev.interactive(), pointsCol = "#D95F02", ylim = NULL, xlim = NULL,
  col = NULL, ...)
}
\arguments{
\item{x}{Fitted "edgwas" object.}

\item{which}{If a subset of the plots is required, specify a subset of the numbers \code{1:5}.}

\item{main}{An overall title for the plot: see \code{\link{title}}}

\item{xlab}{A title for the x axis: see \code{\link{title}}}

\item{ylab}{A title for the y axis: see \code{\link{title}}}

\item{ask}{Logical; if \code{TRUE}, the user is \emph{ask}ed before each plot, see \code{\link{par}(ask=.)}.}

\item{pointsCol}{The color for points corresponding to rho.min and rho.1se. Default is \code{pointsCol = "#D95F02"}.}

\item{ylim}{The y limits of the plot.}

\item{xlim}{The x limits of the plot.}

\item{col}{The color to be used for the curve. Default is \code{col = "grey40"}.}

\item{...}{Other paramters to be passed through to plotting functions.}
}
\description{
Two plots (selectable by which) are currently available: a plot of the mean standard error curve produced by EdGWAS i.e., the MSE and upper and lower standard deviation curves plotted against the values of rho used in the fits, and an adjacency matrix plot for rho.min.
}
\details{
Plots are produced, and nothing is returned.
}
\examples{
N <- 500
q <- 10
set.seed(1)
x <- matrix(sample(0:2, N*q, replace=TRUE), nrow=N, ncol=q)
B <- matrix(0, nrow = q, ncol = q)
B[1, 1:2] <- 10
y <- x \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
###
pc <- edgwas(x, y)
plot(pc, 1)
plot(pc, 2)

}
\seealso{
{\code{\link{edgwas}}}
}
