% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.edgwas.R
\name{predict.edgwas}
\alias{predict.edgwas}
\title{Extract coefficients from a edgwas object}
\usage{
\method{predict}{edgwas}(object, newPS, rho = NULL, type = c("link",
  "coefficients"))
}
\arguments{
\item{object}{Fitted "edgwas" object.}

\item{newPS}{Matrix of new values for x of dimension nobs x nouts (must be polygenic scores - can be generated by \code{\link{ps.edgwas}}) at which predictions are to be made. This argument is not used for \code{type = "coefficients"}.}

\item{rho}{Value(s) of the penalty parameter rho at which predictions are required. Default (NULL) is the entire sequence used to create the model.}

\item{type}{Type of prediction required. Type "link" returns a list of length nouts of matrices of dimension nobs x length(rho) of the fitted values for each response. Type "coefficients" gives the coefficients at the requested values for rho.}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods, this function predicts fitted values and extract coefficients from a fitted "edgwas" object.
}
\examples{

N <- 1000
q <- 10
p <- 100
set.seed(1)
x <- matrix(sample(0:2, N*p, replace=TRUE), nrow=N, ncol=p)
B <- matrix(0, nrow = p, ncol = q)
B[1, 1:2] <- 5
y <- x \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
###
pc <- edgwas(x[-(1:100), ], y[-(1:100), ], scores = FALSE)
ps <- ps.edgwas(x[1:100, ], y[1:100, ])$PS
newy <- predict(pc, newPS = ps)

}
