% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.edgwas.R
\name{cv.edgwas}
\alias{cv.edgwas}
\title{Cross-validation for EdGwas}
\usage{
cv.edgwas(x, y, rho = NULL, nfolds = 10, type.measure = "mse",
  trace = FALSE, ...)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format.}

\item{y}{response matrix, of dimension nobs x nouts. Quantitative for family="gaussian". For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; for a factor, the last level in alphabetical order is the target class). For "binomial" if y is presented as a vector, it will be coerced into a factor.}

\item{rho}{(Non-negative) optional user-supplied rho sequence; default is NULL, and EdGwas chooses its own sequence.}

\item{nfolds}{number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds=3.}

\item{type.measure}{loss to use for cross-validation. Currently one option; the default is type.measure="mse", which uses the mean-squared error.}

\item{...}{Other arguments that can be passed to edgwas.}
}
\value{
Cluster associations. \item{clust}{returns a vector with group memberships}#'
}
\description{
Does k-fold cross-validation for EdGwas, produces a plot, and returns a value for rho.
}
\examples{
# Gaussian
N <- 100 #
q <- 9
p <- 1000 #
set.seed(1)
X <- matrix(sample(0:2, N*p, replace=TRUE), nrow=N, ncol=p)
B <- matrix(0, nrow = p, ncol = q)
B[1:2, 1:5] <- 1
Y <- X \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
###
pc <- cv.edgwas(x = X, y = Y, nfolds = 5)

}
