% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdGwas-package.R
\docType{package}
\name{EdGwas-package}
\alias{EdGwas-package}
\title{EdGwas:}
\description{
A package implementing methods for clustering traits that share some genetic component via polygenic scores (PSs). It fits a sparse precision matrix of the PSs via graphical lasso. The regularisation path is computed for the lasso penalty rho at a grid of values for the regularisation parameter rho. A cross-validation method for tuning the penalty rho is provided.
}
\details{
The EdGwas package provides three categories of important functions:
simple fit, cross-validation and plots.
}
\section{fit}{

The function \code{\link{edgwas}} clusters outcome components (traits) that share some feature (genetic component) via polygenic scores (PSs). It fits a sparse precision matrix via graphical lasso. The regularisation path is computed for the lasso penalty at a grid of values for the regularisation parameter rho.
}

\section{cross-validation}{

The cross-validation function \code{\link{cv.edgwas}} does k-fold cross-validation for edgwas and returns a value for the penalty rho.
}

\section{bootstrap}{

The bootstrap function ...
}

\references{
glasso, glmnet
}
\author{
Ann-Sophie Buchardt \email{asbu@sund.ku.dk}\cr Maintainer: Ann-Sophie Buchardt
\email{asbu@sund.ku.dk}
}
