% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdGwas.R
\docType{package}
\name{EdGwas}
\alias{EdGwas}
\alias{EdGwas-package}
\title{EdGwas: A package implementing methods for clustering traits that share some genetic component via polygenic risk scores (PRSs). It fits a sparse precision matrix via graphical lasso. The regularisation path is computed for the lasso penalty at a grid of values for the regularisation parameter rho.}
\description{
The CSwR package provides three categories of important functions:
simple fit, cross-validation and bootstrap.
}
\section{fit}{

The simple fit function ...
}

\section{cross-validation}{

The cross-validation functions ...
}

\section{bootstrap}{

The bootstrap functions ...
}

\references{
glasso, glmnet
}
\author{
Ann-Sophie Buchardt \email{asbu@sund.ku.dk}\cr Maintainer: Ann-Sophie Buchardt
\email{asbu@sund.ku.dk}
}
