% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgwas.R
\name{edgwas}
\alias{edgwas}
\title{Cluster multiple traits via PRS}
\usage{
edgwas(x, y, rho = NULL, nrho = ifelse(is.null(rho), 20, length(rho)),
  logrho = FALSE, rho.min.ratio = 0.001)
}
\arguments{
\item{x}{Input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format.}

\item{y}{Quantitative response matrix, of dimension nobs x nouts.}

\item{rho}{(Non-negative) regularisation parameter for lasso passed to glasso. rho=0 means no regularisation. Can be a scalar (usual) or a symmetric nouts by nouts matrix, or a vector of length nouts. In the latter case, the penalty matrix has jkth element sqrt(rho[j]*rho[k]).}
}
\value{
Cluster associations. \item{A}{A length(rho) list of matrices A of 0s and 1s, where A_{ij} is equal to 1 iff edges i and j are adjacent and A_{ii} is 0.} \item{rho}{The actual sequence of rho values used.} \item{P}{A length(rho) list of matrices P, the estimated precision matrix (matrix inverse of correlation matrix).}
}
\description{
This function clusters traits that share some genetic component via polygenic risk scores (PRSs). It fits a sparse precision matrix via graphical lasso. The regularisation path is computed for the lasso penalty at a grid of values for the regularisation parameter rho.
}
\details{
...
}
\examples{
# Gaussian
N <- 100
q <- 9
p <- 1000
set.seed(1)
X <- matrix(sample(0:2, N*p, replace=TRUE), nrow=N, ncol=p)
B <- matrix(0, nrow = p, ncol = q)
B[1, 1:2] <- 1
Y <- X \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
###
pc <- edgwas(x = X, y = Y, rho = 0.1)

}
