% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.edgwas.R
\name{predict.cv.edgwas}
\alias{predict.cv.edgwas}
\title{Make predictions from a "cv.edgwas" object}
\usage{
\method{predict}{cv.edgwas}(object, newPS, rho = c("rho.min", "rho.1se"),
  ...)
}
\arguments{
\item{object}{Fitted "cv.edgwas" object.}

\item{newPS}{Matrix of new values for x (must be polygenic scores - can be generated by \code{\link{ps.edgwas}}) at which predictions are to be made. This argument is not used for \code{type = "coefficients"}.}

\item{rho}{Value(s) of the penalty parameter rho at which predictions are required. Default is the value \code{rho = "rho.min"} stored in the object. Alternatively \code{rho = "rho.1se"} can be used. If rho is numeric, it is taken as the value(s) to be used.}

\item{...}{Other arguments to predict.}
}
\value{
The object returned depends on the ... argument which is passed on to the predict method for edgwas objects.
}
\description{
This function makes predictions from a cross-validated edgwas model, using the stored "edgwas.fit" object, and the optimal value chosen for rho.
}
\examples{
N <- 1000
q <- 10
p <- 100
set.seed(1)
x <- matrix(sample(0:2, N*p, replace=TRUE), nrow=N, ncol=p)
B <- matrix(0, nrow = p, ncol = q)
B[1, 1:2] <- 5
y <- x \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
###
pc <- cv.edgwas(x[-(1:100), ], y[-(1:100), ], scores = FALSE)
ps <- ps.edgwas(x[1:100, ], y[1:100, ])$PS
newy <- predict(pc, newPS = ps, rho = "rho.min")

}
