% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.edgwas.R
\name{predict.edgwas}
\alias{predict.edgwas}
\title{Extract coefficients from a edgwas object}
\usage{
\method{predict}{edgwas}(object, newPS, rho = NULL, type = c("link",
  "coefficients"), ...)
}
\arguments{
\item{object}{Fitted "edgwas" object.}

\item{newPS}{Matrix of new values for x of dimension nobs x nouts (must be polygenic scores - can be generated by \code{\link{ps.edgwas}}) at which predictions are to be made. This argument is not used for \code{type = "coefficients"}.}

\item{rho}{Value(s) of the penalty parameter rho at which predictions are required. Default (NULL) is the entire sequence used to create the model.}

\item{type}{Type of prediction required. Type "link" returns a list of length nouts of matrices of dimension nobs x length(rho) of the fitted values for each response. Type "coefficients" gives the coefficients at the requested values for rho.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods, this function predicts fitted values and extract coefficients from a fitted "edgwas" object.
}
\examples{
N <- 1000 #
q <- 10 #
p <- 1000 #
set.seed(1)
# Sample 1
x0 <- matrix(rbinom(n = N*p, size = 2, prob = 0.3), nrow=N, ncol=p)
B <- matrix(0, nrow = p, ncol = q)
B[1, 1:2] <- 2.5
y0 <- x0 \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
beta <- ps.edgwas(x0, y0)$beta
# Sample 2
x <- matrix(rbinom(n = N*p, size = 2, prob = 0.3), nrow=N, ncol=p)
y <- x \%*\% B + matrix(rnorm(N*q), nrow = N, ncol = q)
ps <- x \%*\% beta
###
pc <- edgwas(ps[-(1:100), ], y[-(1:100), ])
newy <- predict(pc, newPS = ps[(1:100), ], rho = pc$rho.min)[[1]]

mean((newy[,1] - y[1:100, 1])^2)
pss <- ps[-(1:100), 1]
lmfit <- lm(y[-(1:100), 1] ~ pss)
newylm <- predict(lmfit, newdata = data.frame(pss = ps[(1:100), 1]))
mean((newylm - y[1:100, 1])^2)

}
